Migration failed
The migration from V3 to V4 worked on Books, but not Authors or Series.  The errors point to something something cloudkit something.   I decided to use what have and start modifying the program with the changes I have in mind.  I'll update the missing data by hand.  I did keep a copy before trying the migration.