checkpoint changes to edit books
I tried to change NewBookView to handle editing of existing books, but when called with a non-nil book the program hangs when I select a book to edit.  I want to save the code as is before I start tweaking it to find out what I (or SwiftUI) is doing wrong.