Book.authors, Author.books, and Series.books from ? to !
The values are initialized to an empty array in the init functions.  Make the Book.authors relationship explicit.  Update authors in the edit view.

Note: I've reached some kind of view complexity that causes a runtime loop if I add any more properties to EditBookView. At least I think it is a loop.  Beachball of death with increasing memory use is what I see.