Change ContentUnavailableView message
Don't tell people to check their search string unless a search is
active.

Also reorder add buttons so the add for the thing being viewed -- books,
authors, and series -- is shown with the others in the three dot menu.
This change is needed because of the new location of the tab bar.